+++ src/wmtk/utils/PartitionMesh.cpp
@@ -2,7 +2,7 @@
 
 namespace wmtk {
 
-tbb::concurrent_vector<int> partition_TriMesh(wmtk::TriMesh m, int num_partition)
+tbb::concurrent_vector<size_t> partition_TriMesh(wmtk::TriMesh m, int num_partition)
 {
     auto f_tuples = m.get_faces();
     Eigen::MatrixXi F(f_tuples.size(), 3);
@@ -13,14 +13,14 @@ tbb::concurrent_vector<int> partition_TriMesh(wmtk::TriMesh m, int num_partition
         F(i, 2) = f_vertice[2].vid();
     }
     auto partitioned_v = partition_mesh_vertices(F, num_partition);
-    tbb::concurrent_vector<int> v_pid(partitioned_v.rows());
+    tbb::concurrent_vector<size_t> v_pid(partitioned_v.rows());
     for (int i = 0; i < partitioned_v.rows(); i++) {
         v_pid[i] = partitioned_v(i, 0);
     }
     return v_pid;
 }
 
-tbb::concurrent_vector<int> partition_TetMesh(wmtk::TetMesh m, int num_partition)
+tbb::concurrent_vector<size_t> partition_TetMesh(wmtk::TetMesh m, int num_partition)
 {
     auto f_tuples = m.get_faces();
     Eigen::MatrixXi F(f_tuples.size(), 3);
@@ -31,7 +31,7 @@ tbb::concurrent_vector<int> partition_TetMesh(wmtk::TetMesh m, int num_partition
         F(i, 2) = e1.switch_edge(m).switch_vertex(m).vid(m);
     }
     auto partitioned_v = partition_mesh_vertices(F, num_partition);
-    tbb::concurrent_vector<int> v_pid(partitioned_v.rows());
+    tbb::concurrent_vector<size_t> v_pid(partitioned_v.rows());
     for (int i = 0; i < partitioned_v.rows(); i++) {
         v_pid[i] = partitioned_v(i, 0);
     }
